fix(pam): address review findings in the access rules table alignment
