+++ bitwarden_license/bit-web/src/app/pam/date/duration-base.pipe.ts
@@ -1,0 +1,39 @@
+import { LOCALE_ID, PipeTransform, inject } from "@angular/core";
+
+import { DurationUnit, pickDurationUnit } from "..";
+
+/**
+ * Shared implementation behind the lease-duration pipes: picks the largest whole unit a
+ * duration divides into (via {@link pickDurationUnit}) and renders it with
+ * `Intl.NumberFormat`'s `style: "unit"`, so both the unit label and its plural form follow
+ * the active locale rather than a hand-rolled English string.
+ *
+ * Subclasses supply only the `unitDisplay` style; unit selection, the per-unit formatter
+ * cache, and the locale source live here so the compact and spelled-out renderings cannot
+ * drift apart.
+ */
+export abstract class DurationBasePipe implements PipeTransform {
+  private readonly locale = inject(LOCALE_ID);
+  private readonly formatters = new Map<DurationUnit, Intl.NumberFormat>();
+
+  /** `Intl.NumberFormat`'s `unitDisplay` style this pipe renders with. */
+  protected abstract readonly unitDisplay: Intl.NumberFormatOptions["unitDisplay"];
+
+  transform(seconds: number): string {
+    const { value, unit } = pickDurationUnit(seconds);
+    return this.formatterFor(unit).format(value);
+  }
+
+  private formatterFor(unit: DurationUnit): Intl.NumberFormat {
+    let formatter = this.formatters.get(unit);
+    if (!formatter) {
+      formatter = new Intl.NumberFormat(this.locale, {
+        style: "unit",
+        unit,
+        unitDisplay: this.unitDisplay,
+      });
+      this.formatters.set(unit, formatter);
+    }
+    return formatter;
+  }
+}
